+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Show Folder status including Shared status, source project and date of sharing
   - Show tags for Folders and Objects inside them
   - Change routes from Folder view back to the equivalent Folder tab
+- (GH #751) Add modal for copying folder
 - (GH #746) Move container and object listings to csc-ui components
 - (GH #746) Migrate buefy search to CSC styling, csc-ui component for search not an option
 - (GH #714) Add modal for sharing folder with other projects